updated ruleset to reject if disallowedTags were found
