+++ src/lib/schema/html-input.ts
@@ -1,0 +1,49 @@
+/// Adapted from https://github.com/dekatron/joi-html-input
+
+import joi from 'joi';
+import sanitizeHtml from 'sanitize-html';
+
+const displayString: (string) => string = (input) => {
+    const sanitizeConfig = {
+        allowedTags: [],
+        allowedAttributes: {},
+    };
+    return sanitizeHtml(input, sanitizeConfig);
+};
+type HtmlInputOptions = sanitizeHtml.IOptions;
+
+const htmlInputExtension: joi.Extension = {
+    type: 'htmlInput',
+    base: joi.string(),
+    messages: {
+        'htmlInput.invalidTags': '{{#label}} input must not contain html tags',
+    },
+    rules: {
+        allowedTags: {
+            method(options: HtmlInputOptions) {
+                return this.$_addRule({
+                    name: 'allowedTags',
+                    args: { options },
+                });
+            },
+            args: [
+                {
+                    name: 'options',
+                    assert: joi.object().keys({
+                        allowedTags: joi.array().items(joi.string()),
+                        allowedAttributes: joi.object(),
+                    }),
+                },
+            ],
+            validate(value, helpers, args) {
+                const sanitized = sanitizeHtml(value, args.options);
+                if (sanitized !== value) {
+                    return helpers.error('htmlInput.invalidTags');
+                }
+                return value;
+            },
+        },
+    },
+};
+
+export default htmlInputExtension;
